+++ motan-core/src/main/java/com/weibo/api/motan/filter/AccessStatisticFilter.java
@@ -18,10 +18,7 @@ package com.weibo.api.motan.filter;
 
 import com.weibo.api.motan.common.MotanConstants;
 import com.weibo.api.motan.core.extension.SpiMeta;
-import com.weibo.api.motan.rpc.Caller;
-import com.weibo.api.motan.rpc.Provider;
-import com.weibo.api.motan.rpc.Request;
-import com.weibo.api.motan.rpc.Response;
+import com.weibo.api.motan.rpc.*;
 import com.weibo.api.motan.util.ExceptionUtil;
 import com.weibo.api.motan.util.MotanFrameworkUtil;
 import com.weibo.api.motan.util.StatsUtil;
@@ -38,46 +35,48 @@ public class AccessStatisticFilter implements Filter {
     private static final String RPC_SERVICE = "rpc_service";
 
     @Override
-    public Response filter(Caller<?> caller, Request request) {
+    public Response filter(final Caller<?> caller, final Request request) {
         long start = System.currentTimeMillis();
         AccessStatus accessStatus = AccessStatus.NORMAL;
-        boolean specialException = true;
-        long bizProcessTime = 0;
+        final long bizProcessTime;
+        Response response = null;
 
         try {
-            Response response = caller.call(request);
-
-            if (response == null) {
-                accessStatus = AccessStatus.OTHER_EXCEPTION;
-            } else {
-                if (response.getException() != null) {
-                    if (ExceptionUtil.isBizException(response.getException())) {
-                        accessStatus = AccessStatus.BIZ_EXCEPTION;
-                    } else {
-                        accessStatus = AccessStatus.OTHER_EXCEPTION;
-                    }
+            response = caller.call(request);
+            if (response != null && response.getException() != null) {
+                if (ExceptionUtil.isBizException(response.getException())) {
+                    accessStatus = AccessStatus.BIZ_EXCEPTION;
+                } else {
+                    accessStatus = AccessStatus.OTHER_EXCEPTION;
                 }
-
-                specialException = false;
-                bizProcessTime = response.getProcessTime();
             }
-
             return response;
         } finally {
             long end = System.currentTimeMillis();
 
-            if (specialException) {
+            if (response == null) {
                 accessStatus = AccessStatus.OTHER_EXCEPTION;
                 bizProcessTime = end - start;
+            } else {
+                bizProcessTime = response.getProcessTime();
             }
 
-            String statName =
-                    caller.getUrl().getProtocol() + MotanConstants.PROTOCOL_SEPARATOR + MotanFrameworkUtil.getGroupMethodString(request);
+            final String statName = caller.getUrl().getProtocol() + MotanConstants.PROTOCOL_SEPARATOR + MotanFrameworkUtil.getGroupMethodString(request);
             if (caller instanceof Provider) {
                 StatsUtil.accessStatistic(statName, APPLICATION_STATISTIC, RPC_SERVICE, end, end - start, bizProcessTime, accessStatus);
+                if (request instanceof TraceableRequest) {
+                    final AccessStatus finalAccessStatus = accessStatus;
+                    ((TraceableRequest) request).addFinishCallback(new Runnable() {
+                        @Override
+                        public void run() {
+                            long requestEnd = ((TraceableRequest) request).getEndTime();
+                            long requestStart = ((TraceableRequest) request).getStartTime();
+                            StatsUtil.accessStatistic(statName + "_WHOLE", caller.getUrl().getApplication(), caller.getUrl().getModule(), requestEnd, requestEnd - requestStart, bizProcessTime, finalAccessStatus);
+                        }
+                    }, null);
+                }
             }
             StatsUtil.accessStatistic(statName, caller.getUrl().getApplication(), caller.getUrl().getModule(), end, end - start, bizProcessTime, accessStatus);
-
         }
     }
 }