+++ motan-core/src/main/java/com/weibo/api/motan/registry/support/AbstractRegistry.java
@@ -43,8 +43,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 public abstract class AbstractRegistry implements Registry {
 
-    private ConcurrentHashMap<URL, Map<String, List<URL>>> subscribedCategoryResponses =
-            new ConcurrentHashMap<URL, Map<String, List<URL>>>();
+    private ConcurrentHashMap<URL, Map<String, List<URL>>> subscribedCategoryResponses = new ConcurrentHashMap<>();
 
     private URL registryUrl;
     private Set<URL> registeredServiceUrls = new ConcurrentHashSet<URL>();