+++ motan-transport-netty4/src/test/java/com/weibo/api/motan/transport/netty4/NettyServerTest.java
@@ -6,6 +6,7 @@ import com.weibo.api.motan.rpc.Request;
 import com.weibo.api.motan.rpc.URL;
 import com.weibo.api.motan.transport.Channel;
 import com.weibo.api.motan.transport.MessageHandler;
+import com.weibo.api.motan.util.StatsUtil;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
@@ -70,6 +71,37 @@ public class NettyServerTest {
         nettyClient2.close();
         Thread.sleep(100);
         Assert.assertEquals(0, nettyServer.channelManage.getChannels().size());
+    }
+
+    @Test
+    public void testCallbacks() throws InterruptedException {
+        int minClientConnection = 2;
+        int maxServerConnection = 2;
+        url.addParameter(URLParamType.minClientConnection.getName(), String.valueOf(minClientConnection));
+        url.addParameter(URLParamType.maxServerConnection.getName(), String.valueOf(maxServerConnection));
+        url.addParameter(URLParamType.requestTimeout.getName(), "10000");
+        nettyServer = new NettyServer(url, new MessageHandler() {
+            @Override
+            public Object handle(Channel channel, Object message) {
+                Request request = (Request) message;
+                DefaultResponse response = new DefaultResponse();
+                response.setRequestId(request.getRequestId());
+                response.setValue("method: " + request.getMethodName() + " requestId: " + request.getRequestId());
+                return response;
+            }
+        });
+        nettyServer.open();
+        Assert.assertEquals(0, nettyServer.channelManage.getChannels().size());
+
+        NettyClient nettyClient = new NettyClient(url);
+        nettyClient.open();
+        Thread.sleep(100);
+        nettyClient.close();
+
+        Assert.assertEquals(minClientConnection + 1, StatsUtil.getStatisticCallbacks().size());
+        nettyServer.close();
+        Thread.sleep(100);
+        Assert.assertEquals(0, StatsUtil.getStatisticCallbacks().size());
     }
 
 }